Prepared for static scraper integration
