+++ main.py
@@ -7,7 +7,7 @@ import psutil
 from models import URL, ScraperStatusTypeEnum, ContentSourceTypeEnum
 from database import SessionLocal
 from datetime import datetime
-from scrapers.scraper import Scraper
+from scraper import Scraper
 from custom_logger import get_logger
