+++ models.py
@@ -26,7 +26,6 @@ class URL(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     url = Column(String(255), nullable=False)
-    description = Column(Text, nullable=True)
     last_scraped_date = Column(DateTime, nullable=True)
     is_dead = Column(Boolean, default=False)
     content_hash = Column(Text, nullable=True)
