+++ scraper.py
@@ -39,7 +39,7 @@ class Scraper:
 
         driver = webdriver.Firefox(options=options)
 
-        with open('../def.js', 'r', encoding='utf-8') as script_file:
+        with open('def.js', 'r', encoding='utf-8') as script_file:
             driver.execute_script(script_file.read())
 
         # driver.set_page_load_timeout(self.__time_out)
@@ -102,7 +102,7 @@ async def main(targets: list[str]):
 
 if __name__ == '__main__':
     import json
-    with open('../dataset.json', 'r', encoding='utf-8') as dataset:
+    with open('dataset.json', 'r', encoding='utf-8') as dataset:
         data = json.loads(dataset.read())
 
     targets: list[str] = [url for cve in data for url in cve['urls']]
