--- scrapers/death_prejudicer.py
@@ -1,123 +1,0 @@
-import asyncio
-import os
-
-import aiofiles
-import aiohttp
-import threading
-from models import URL, ScraperStatusTypeEnum
-from database import SessionLocal, engine
-from sqlalchemy.orm import sessionmaker
-from custom_logger import get_logger
-from hashlib import sha256
-
-logger = get_logger('DeathPrejudice')
-
-Session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-
-async def check_url(http_session: aiohttp.ClientSession, url, url_id, cve_id, results):
-    try:
-        async with http_session.get(url) as response:
-            death_prejudice = response.status not in [200, 202, 403, 401]
-            if response.status == 200:
-                if response.headers.get("Content-Type", None) == 'application/pdf':
-                    pdf_content = await response.read()
-
-                    folder_path = f'content-context/{cve_id}'
-                    os.makedirs(folder_path, exist_ok=True)
-                    hash_context = sha256(pdf_content).hexdigest()
-
-                    file_path = f'{folder_path}/{hash_context}.pdf'
-                    async with aiofiles.open(file_path, 'wb') as pdf:
-                        await pdf.write(pdf_content)
-                else:
-                    html_content = await response.text()
-
-                    folder_path = f'content-context/{cve_id}'
-                    os.makedirs(folder_path, exist_ok=True)
-                    hash_context = sha256(html_content.encode()).hexdigest()
-                    file_path = f'{folder_path}/{hash_context}.html'
-                    async with aiofiles.open(file_path, 'wb') as pdf:
-                        await pdf.write(html_content.encode())
-            results.append({
-                'id': url_id,
-                'status': response.status,
-                'death_prejudice': death_prejudice,
-
-            })
-
-    except Exception as e:
-        logger.exception(e)
-        results.append(
-            {'id': url_id, 'status': 0, 'death_prejudice': True, 'scraper_status': ScraperStatusTypeEnum.PROBLEM}
-        )
-        pass
-
-
-async def process_urls(urls, http_session: aiohttp.ClientSession):
-    results = []
-    tasks = []
-    for url, url_id, scraper_status, cve_id in urls:
-        tasks.append(check_url(http_session, url, url_id, cve_id, results))
-    await asyncio.gather(*tasks)
-    return results
-
-
-def bulk_update_db(db, results):
-    update_data = []
-
-    for result in results:
-        update_data.append({
-            'id': result['id'],
-            'http_status': result['status'],
-            'death_prejudice': result.get('death_prejudice', False),
-            'scraper_status': ScraperStatusTypeEnum.READY
-        })
-
-    db.bulk_update_mappings(URL, update_data)
-    db.commit()
-
-
-def thread_worker(urls_chunk):
-    db = Session()
-
-    async def run_async_tasks():
-        async with aiohttp.ClientSession() as http_session:
-            results = await process_urls(urls_chunk, http_session)
-            bulk_update_db(db, results)
-
-    asyncio.run(run_async_tasks())
-    db.close()
-    print(f"Processed {len(urls_chunk)} URLs.")
-
-
-def main():
-    db = SessionLocal()
-
-    urls_with_zero_status = db.query(URL).filter(URL.scraper_status != ScraperStatusTypeEnum.READY).all()
-
-    targets = [(url.url, url.id, url.scraper_status, url.cve_id) for url in urls_with_zero_status]
-
-    chunk_size = 40
-    chunks = [targets[i:i + chunk_size] for i in range(0, len(targets), chunk_size)]
-    print('Chunks loaded')
-
-    thread_count = 4
-    threads = []
-    for i in range(0, len(chunks), thread_count):
-        threads_to_run = chunks[i:i + thread_count]
-        for chunk in threads_to_run:
-            thread = threading.Thread(target=thread_worker, args=(chunk,))
-            threads.append(thread)
-            thread.start()
-
-    for thread in threads:
-        thread.join()
-
-    db.close()
-    print('All URLs processed and results saved to the database.')
-
-
-if __name__ == "__main__":
-    # pdf test case https://9443417.fs1.hubspotusercontent-na1.net/hubfs/9443417/Security%20Advisories/AMI-SA-2023005.pdf
-    main()
