+++ scrapers/scrape_with_static.py
@@ -1,7 +1,114 @@
 import asyncio
+from datetime import datetime
 import aiohttp
 import threading
+from models import URL, ScraperStatusTypeEnum
+from database import SessionLocal, engine
+from sqlalchemy.orm import sessionmaker
 from custom_logger import get_logger
+from scrapers.utilities import async_file_typer
 
 
-logger = get_logger('Static Scraper')
+logger = get_logger('DeathPrejudice')
+Session = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+
+async def check_url(http_session: aiohttp.ClientSession, results: list, **target):
+    try:
+        url = target.get('url')
+        cve_id = target.get('cve_id')
+        http_status: int = 0
+        async with http_session.get(url) as response:
+            death_prejudice = response.status not in [200, 202, 403, 401]
+            http_status = response.status
+            if response.status == 200:
+                folder_path = f'content-context/{cve_id}'
+
+                if response.headers.get("Content-Type", None) == 'application/pdf':
+                    pdf_content = await response.read()
+                    tf = await async_file_typer(content=pdf_content, folder_path=folder_path, extension='pdf')
+                else:
+                    html_content = await response.text()
+                    tf = await async_file_typer(content=html_content.encode(), folder_path=folder_path, extension='html')
+                results.append({
+                    'id': target.get('id'),
+                    'http_status': response.status,
+                    'death_prejudice': death_prejudice,
+                    'content_hash': tf.hash,
+                    'content_path': tf.path,
+                    'last_scraped_date': datetime.now(),
+                    'scraper_status': ScraperStatusTypeEnum.DONE
+                })
+
+    except Exception as e:
+        logger.exception(e)
+        results.append({
+            'id': target.get('id'),
+            'status': 0,
+            'death_prejudice': True,
+            'scraper_status': ScraperStatusTypeEnum.PROBLEM,
+            'http_status': http_status
+        })
+        pass
+
+
+async def process_urls(urls, http_session: aiohttp.ClientSession):
+    results = []
+    tasks = []
+    for url in urls:
+        tasks.append(check_url(http_session=http_session, results=results, **url))
+    await asyncio.gather(*tasks)
+    return results
+
+
+def bulk_update_db(db, results):
+    db.bulk_update_mappings(URL, results)
+    db.commit()
+
+
+def thread_worker(urls_chunk):
+    db = Session()
+
+    async def run_async_tasks():
+        async with aiohttp.ClientSession() as http_session:
+            results = await process_urls(urls_chunk, http_session)
+            bulk_update_db(db, results)
+
+    asyncio.run(run_async_tasks())
+    db.close()
+    print(f"Processed {len(urls_chunk)} URLs.")
+
+
+def scrape_with_static():
+    db = SessionLocal()
+    target_list = db.query(URL).filter(URL.scraper_status != ScraperStatusTypeEnum.READY).all()
+
+    # target_list = db.query(URL).filter(
+    #     URL.scraper_status != ScraperStatusTypeEnum.READY,
+    #     URL.url.like('%.pdf')
+    # ).all()
+
+    targets = [url.__dict__ for url in target_list[:20]]
+
+    chunk_size = 40
+    chunks = [targets[i:i + chunk_size] for i in range(0, len(targets), chunk_size)]
+    print('Chunks loaded')
+
+    thread_count = 4
+    threads = []
+    for i in range(0, len(chunks), thread_count):
+        threads_to_run = chunks[i:i + thread_count]
+        for chunk in threads_to_run:
+            thread = threading.Thread(target=thread_worker, args=(chunk,))
+            threads.append(thread)
+            thread.start()
+
+    for thread in threads:
+        thread.join()
+
+    db.close()
+    print('All URLs processed and results saved to the database.')
+
+
+if __name__ == "__main__":
+    scrape_with_static()
