+++ scrapers/utilities/async_file_typer.py
@@ -1,0 +1,20 @@
+import os
+import aiofiles
+from typing import ByteString, AnyStr
+from hashlib import sha256
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class AsyncTypedFile:
+    path: str
+    hash: hash
+
+
+async def async_file_typer(content: ByteString, folder_path: AnyStr, extension: AnyStr) -> AsyncTypedFile:
+    os.makedirs(folder_path, exist_ok=True)
+    hash_context = sha256(content).hexdigest()
+    file_path = f'{folder_path}/{hash_context}.{extension}'
+    async with aiofiles.open(file_path, 'wb') as _content:
+        await _content.write(content)
+    return AsyncTypedFile(path=file_path, hash=hash_context)
